BAU: Pull most benchmark logic into an abstract class
Since we've abstracted the CardInformationStore, it seems logical that we
should also abstract the benchmarking of it, in case we want more than one
implementation in future.
